add u16 pin

+++ src/mlib/lib_pins.rs
@@ -1,19 +1,19 @@
 pub mod pins_wiringpi {
-    pub const BEEP_PIN: u8 = 0;
-    pub const POW_PWM_L_PIN: u8 = 1;
-    pub const POW_IN2_L_PIN: u8 = 2;
-    pub const POW_IN1_L_PIN: u8 = 3;
-    pub const POW_PWM_R_PIN: u8 = 4;
-    pub const POW_IN2_R_PIN: u8 = 5;
-    pub const POW_IN1_R_PIN: u8 = 6;
-    pub const BTN_PIN: u8 = 7;
-    pub const RED_PIN: u8 = 21;
-    pub const GREEN_PIN: u8 = 22;
-    pub const LEFT_TCR_PIN: u8 = 25;
-    pub const MID_TCR_PIN: u8 = 24;
-    pub const RIGHT_TCR_PIN: u8 = 23;
-    pub const LEFT_OA_PIN: u8 = 26;
-    pub const RIGHT_OA_PIN: u8 = 27;
-    pub const TRIG_SEND: u8 = 28;
-    pub const TRIG_RECV: u8 = 29;
+    pub const BEEP_PIN: u16 = 0;
+    pub const POW_PWM_L_PIN: u16 = 1;
+    pub const POW_IN2_L_PIN: u16 = 2;
+    pub const POW_IN1_L_PIN: u16 = 3;
+    pub const POW_PWM_R_PIN: u16 = 4;
+    pub const POW_IN2_R_PIN: u16 = 5;
+    pub const POW_IN1_R_PIN: u16 = 6;
+    pub const BTN_PIN: u16 = 7;
+    pub const RED_PIN: u16 = 21;
+    pub const GREEN_PIN: u16 = 22;
+    pub const LEFT_TCR_PIN: u16 = 25;
+    pub const MID_TCR_PIN: u16 = 24;
+    pub const RIGHT_TCR_PIN: u16 = 23;
+    pub const LEFT_OA_PIN: u16 = 26;
+    pub const RIGHT_OA_PIN: u16 = 27;
+    pub const TRIG_SEND: u16 = 28;
+    pub const TRIG_RECV: u16 = 29;
 }
